[web] Move insertion pages to a separate directory
