+++ web/src/index.tsx
@@ -13,18 +13,18 @@ import "jquery/dist/jquery.min.js";
 import "bootstrap/dist/js/bootstrap.min.js";
 
 import Homepage from "./routes/Homepage/Homepage";
-import NewAssignment from "./routes/NewAssignment/NewAssignment";
-import NewGroup from "./routes/NewGroup/NewGroup";
-import NewUser from "./routes/NewUser/NewUser";
-import NewPermission from "./routes/NewPermission/NewPermission";
-import NewRole from "./routes/NewRole/NewRole";
-import NewDelayedAssignment from "./routes/NewDelayedAssignment/NewDelayedAssignment";
-import NewTemporaryRole from "./routes/NewTemporaryRole/NewTemporaryRole";
-import NewPublicGroup from "./routes/NewPublicGroup/NewPublicGroup";
-import NewUserHasRole from "./routes/NewUserHasRole/NewUserHasRole";
-import NewUserHasCompletedAssignment from "./routes/NewUserHasCompletedAssignment/NewUserHasCompletedAssignment";
-import NewAssignmentCreatedByUser from "./routes/NewAssignmentCreatedByUser/NewAssignmentCreatedByUser";
-import NewUserMemberOfGroup from "./routes/NewUserMemberOfGroup/NewUserMemberOfGroup";
+import NewAssignment from "./routes/new/NewAssignment/NewAssignment";
+import NewGroup from "./routes/new/NewGroup/NewGroup";
+import NewUser from "./routes/new/NewUser/NewUser";
+import NewPermission from "./routes/new/NewPermission/NewPermission";
+import NewRole from "./routes/new/NewRole/NewRole";
+import NewDelayedAssignment from "./routes/new/NewDelayedAssignment/NewDelayedAssignment";
+import NewTemporaryRole from "./routes/new/NewTemporaryRole/NewTemporaryRole";
+import NewPublicGroup from "./routes/new/NewPublicGroup/NewPublicGroup";
+import NewUserHasRole from "./routes/new/NewUserHasRole/NewUserHasRole";
+import NewUserHasCompletedAssignment from "./routes/new/NewUserHasCompletedAssignment/NewUserHasCompletedAssignment";
+import NewAssignmentCreatedByUser from "./routes/new/NewAssignmentCreatedByUser/NewAssignmentCreatedByUser";
+import NewUserMemberOfGroup from "./routes/new/NewUserMemberOfGroup/NewUserMemberOfGroup";
 import { ButtonGroup } from "react-bootstrap";
 import Button from "react-bootstrap/Button";
 import NotFound from "./routes/NotFound/NotFound";
