+++ web/src/routes/new/NewAssignment/NewAssignment.tsx
@@ -2,10 +2,10 @@ import React from "react";
 import { Form, FormControl, FormGroup, FormLabel } from "react-bootstrap";
 import Button from "react-bootstrap/Button";
 import { useForm } from "react-hook-form";
-import ToastHelper from "../../components/ToastHelper";
-import { AssignmentRequest } from "../../openapi";
-import { Api } from "../../index";
-import { ErrorToast, SuccessToast } from "../../components/MyToasts";
+import ToastHelper from "../../../components/ToastHelper";
+import { AssignmentRequest } from "../../../openapi";
+import { Api } from "../../../index";
+import { ErrorToast, SuccessToast } from "../../../components/MyToasts";
 
 interface NewAssignmentProps {}
