+++ web/src/routes/new/NewAssignmentCreatedByUser/NewAssignmentCreatedByUser.module.scss
