+++ web/src/routes/new/NewPermission/NewPermission.tsx
@@ -1,29 +1,30 @@
 import React from "react";
 import { Form, FormControl, FormGroup, FormLabel } from "react-bootstrap";
 import Button from "react-bootstrap/Button";
-import { Api } from "../../index";
+import ToastHelper from "../../../components/ToastHelper";
+import { PermissionRequest } from "../../../openapi";
+import { Api } from "../../../index";
 import { useForm } from "react-hook-form";
-import { GroupRequest } from "../../openapi";
-import ToastHelper from "../../components/ToastHelper";
-import { ErrorToast, SuccessToast } from "../../components/MyToasts";
+import { ErrorToast, SuccessToast } from "../../../components/MyToasts";
 
-interface NewGroupProps {}
+interface NewPermissionProps {}
 
 const helper = new ToastHelper();
 
-async function submitForm(groupRequest: GroupRequest) {
-  await helper.takeoverPromise(Api.groupsAddPost(groupRequest));
+async function submitForm(data: PermissionRequest) {
+  await helper.takeoverPromise(Api.permissionsAddPost(data));
 }
 
-function NewGroup() {
+function NewPermission() {
   const {
     register,
     handleSubmit,
     formState: { errors },
-  } = useForm<GroupRequest>();
+  } = useForm<PermissionRequest>();
+
   return (
     <Form onSubmit={handleSubmit((data) => submitForm(data))}>
-      <h2>Create New Group</h2>
+      <h2>Create New Permission</h2>
       <FormGroup>
         <FormLabel>Name</FormLabel>
         <FormControl type="text" {...register("name", { required: true })} />
@@ -32,9 +33,12 @@ function NewGroup() {
         <FormLabel>Description</FormLabel>
         <Form.Control
           as="textarea"
-          id="description"
-          {...register("description")}
+          {...register("description", { required: true })}
         />
+      </FormGroup>
+      <FormGroup>
+        <FormLabel>Type</FormLabel>
+        <FormControl type="number" {...register("type", { required: true })} />
       </FormGroup>
       <FormGroup>
         <Button type="submit" className={"mt-3"}>
@@ -50,4 +54,4 @@ function NewGroup() {
   );
 }
 
-export default NewGroup;
+export default NewPermission;
