+++ web/src/routes/new/NewUserHasCompletedAssignment/NewUserHasCompletedAssignment.module.scss
