+++ web/src/routes/new/NewUserHasRole/NewUserHasRole.tsx
@@ -1,30 +1,29 @@
 import React, { useEffect, useState } from "react";
-import ToastHelper from "../../components/ToastHelper";
-import { PublicGroup } from "../../openapi";
-import { Api } from "../../index";
-import { useForm } from "react-hook-form";
 import { Form, FormGroup, FormLabel, FormSelect } from "react-bootstrap";
 import Button from "react-bootstrap/Button";
-import { ErrorToast, SuccessToast } from "../../components/MyToasts";
+import { Api } from "../../../index";
+import ToastHelper from "../../../components/ToastHelper";
+import { ErrorToast, SuccessToast } from "../../../components/MyToasts";
+import { useForm } from "react-hook-form";
+import { UserHasRole } from "../../../openapi";
 
-interface NewPublicGroupProps {}
+interface NewUserHasRoleProps {}
 
 const helper = new ToastHelper();
 
-async function submitForm(data: PublicGroup) {
-  await helper.takeoverPromise(Api.publicGroupsAddPost(data));
+async function submitForm(data: UserHasRole) {
+  await helper.takeoverPromise(Api.userHasRoleAddPost(data));
 }
 
-function NewPublicGroup() {
+function NewUserHasRole() {
   const {
     register,
     handleSubmit,
     formState: { errors },
-  } = useForm<PublicGroup>();
+  } = useForm<UserHasRole>();
 
   const [roles, updateRoles] = useState([] as JSX.Element[]);
 
-  // load all groups
   useEffect(() => {
     Api.rolesAllGet().then((res) => {
       const objs = res.data.response.map((role) => (
@@ -36,32 +35,30 @@ function NewPublicGroup() {
     });
   }, []);
 
-  const [groups, updateGroups] = useState([] as JSX.Element[]);
-
-  // load all groups
+  const [users, updateUsers] = useState([] as JSX.Element[]);
   useEffect(() => {
-    Api.groupsAllGet().then((res) => {
-      const objs = res.data.response.map((group) => (
-        <option value={group.group_id} key={"group_" + group.group_id}>
-          {group.name}
+    Api.usersAllGet().then((res) => {
+      const objs = res.data.response.map((user) => (
+        <option value={user.user_id} key={"user_" + user.user_id}>
+          {user.first_name + " " + user.last_name}
         </option>
       ));
-      updateGroups(objs);
+      updateUsers(objs);
     });
   }, []);
 
   return (
     <Form onSubmit={handleSubmit((data) => submitForm(data))}>
-      <h2>Create New Public Group</h2>
+      <h2>Create New User Has Role Relation</h2>
       <FormGroup>
-        <FormLabel>Group</FormLabel>
-        <FormSelect {...register("group_id", { required: true })}>
-          {groups}
+        <FormLabel>User</FormLabel>
+        <FormSelect {...register("user_id", { required: true })}>
+          {users}
         </FormSelect>
       </FormGroup>
       <FormGroup>
         <FormLabel>Role</FormLabel>
-        <FormSelect {...register("default_role_id", { required: true })}>
+        <FormSelect {...register("role_id", { required: true })}>
           {roles}
         </FormSelect>
       </FormGroup>
@@ -79,4 +76,4 @@ function NewPublicGroup() {
   );
 }
 
-export default NewPublicGroup;
+export default NewUserHasRole;
